Move scolapasta-hex tests into their own files
They were getting large.
